Move AdminDashboard to pages\technical\
